Change animal age from integer to floating point number

+++ AnimalShelter/Models/Animal.cs
@@ -9,7 +9,7 @@ namespace AnimalShelter.Models
         public string Color { get; set; }
         public string Size { get; set; }
         public bool SpayedOrNeutered { get; set; }
-        public int Age { get; set; }
+        public float Age { get; set; }
         public string Gender { get; set; }
         public string Description { get; set; }
     }
